fix: 방명록 남기기 페이지 로그인 요구 비활성화
- 방명록 남기기 페이지 로그인 요구 비활성화
- 방명록을 남긴 다음 방문자는 상세 페이지를 볼 수 있음
  - 비회원은 메뉴바가 보이지 않음
