+++ src/app/guest-book/create/_hook/useGuestBookLink.ts
@@ -1,13 +1,11 @@
 import api from '@/api';
-import useAuth from '@/contexts/auth.context';
 import { showToast } from '@/libs/utils';
 import TImageFile from '@/types/image.file';
 import { useMutation } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 const useGuestBookLink = () => {
-  const { signedIn, loading } = useAuth();
   const [visitorName, setVisitorName] = useState('');
   const [file, setFile] = useState<TImageFile | null>(null);
   const router = useRouter();
@@ -37,13 +35,6 @@ const useGuestBookLink = () => {
       showToast.error('방명록 링크 생성이 실패했습니다.');
     },
   });
-
-  useEffect(() => {
-    if (!signedIn && !loading) {
-      showToast.error('로그인 후 이용해주세요!');
-      router.push('/users');
-    }
-  }, [signedIn, router, loading]);
 
   return { visitorName, setVisitorName, file, setFile, createLink };
 };
